Implementasi Silcing Coding Test
Seluruh implementasi slicing telah selesai di buat dengan membagi halaman Main menjadi lima base component.
