+++ package.json
@@ -33,7 +33,7 @@
 		"style-loader": "^3.3.2",
 		"svg-url-loader": "^8.0.0",
 		"webpack": "^5.78.0",
-		"webpack-cli": "^5.0.1",
+		"webpack-cli": "^5.1.4",
 		"webpack-dev-server": "^4.13.2"
 	},
 	"dependencies": {
