+++ src/Pages/Main/Title.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import styles from "./Main.module.css";
+
+const Title = () => {
+  return (
+    <div className={styles.titleContent}>
+      <div className={styles.textTitle}>Lorem Ipsum Dolor Sit</div>
+      <div className={styles.textDesc}>Lorem Ipsum Dolor Sit</div>
+    </div>
+  );
+};
+
+export default Title;
